[fix] Login, Signup Coordinator 구조 재수정 #19
