+++ FindTown/FindTown/Presentation/AuthScene/Flows/LoginCoordinatorDelegate.swift
@@ -1,0 +1,13 @@
+//
+//  LoginCoordinatorDelegate.swift
+//  FindTown
+//
+//  Created by 김성훈 on 2023/01/06.
+//
+
+import Foundation
+
+protocol LoginCoordinatorDelegate {
+    func goToNickname()
+    func goToTabBar()
+}
