+++ FindTown/FindTown/Presentation/AuthScene/Login/LoginViewModel.swift
@@ -11,20 +11,27 @@ import FindTownCore
 
 protocol LoginViewModelType {
     func goToNicnName()
+    func goToTabBar()
 }
 
 final class LoginViewModel: BaseViewModel {
-    let delegate: SignupCoordinatorDelegate
+    
+    let delegate: LoginCoordinatorDelegate
     
     init(
-        delegate: SignupCoordinatorDelegate
+        delegate: LoginCoordinatorDelegate
     ) {
         self.delegate = delegate
     }
 }
 
 extension LoginViewModel: LoginViewModelType {
+    
     func goToNicnName() {
         delegate.goToNickname()
+    }
+    
+    func goToTabBar() {
+        delegate.goToTabBar()
     }
 }
